DSP::Synapse ctor: 97.39 -> 97.83, the three sample-count conversions are float -> unsigned
The image converts with fctidz and stores the low word (0x82E47190 stfd
0x58 / lwz 0x5c for the ring-buffer length, 0x82E471A0 / 0x82E471AC
fctidz + stfiwx straight into mDefaultPitch and 0x20).  That is MSVC's
float -> unsigned int lowering; the (int) casts we had in front of the
unsigned members produced fctiwz, which also saturates differently on an
out-of-range sample rate.  Members are declared unsigned int already.

+++ src/system/synth_xbox/Synapse_dsp.cpp
@@ -175,9 +175,9 @@ Synapse::Synapse(float sampleRate) : mDetectionInterval(64), mTargetPitch(sample
     float prod2 = mTargetPitch * 0.0015384615f;
     float prod3 = mTargetPitch * 0.016666668f;
     unsigned int inputLen =
-        (unsigned int)(int)(prod1 + (prod1 >= 0.0f ? 0.5f : -0.5f)) & ~3u;
-    mDefaultPitch = (int)(prod2 + (prod2 >= 0.0f ? 0.5f : -0.5f));
-    mField_0x20 = (int)(prod3 + (prod3 >= 0.0f ? 0.5f : -0.5f));
+        (unsigned int)(prod1 + (prod1 >= 0.0f ? 0.5f : -0.5f)) & ~3u;
+    mDefaultPitch = (unsigned int)(prod2 + (prod2 >= 0.0f ? 0.5f : -0.5f));
+    mField_0x20 = (unsigned int)(prod3 + (prod3 >= 0.0f ? 0.5f : -0.5f));
 
     // The fill values are unnamed temporaries; the target reuses ONE stack
     // slot for all of them (and for the ChannelBuffer prototype / coeffs),
